requirements.txt
streamlit
deepface
spotipy
opencv-python-headless
tensorflow
Pillow
